+++ src/models/gvrp_models/cplex/matheus_model/energy_ub_constraint.cpp
@@ -18,7 +18,7 @@ Energy_ub_constraint::Energy_ub_constraint (Matheus_model& matheus_model) : Extr
 
 void Energy_ub_constraint::add () {
   IloConstraint c;
-  for (size_t i = 1; i < matheus_model.c0.size(); ++i){
+  for (int i = 1; i < matheus_model.c0.size(); ++i){
     /*
     c = IloConstraint (matheus_models.e[i - 1] <= calculateCustomerMaxRequiredFuel(matheus_models.instance, *matheus_models.gvrp_afs_tree, *matheus_models.c0[i]));
     c.setName("Energy level UB");
