+++ src/models/gvrp_models/cplex/matheus_model/invalid_edge_preprocessing.cpp
@@ -33,21 +33,21 @@ void Invalid_edge_preprocessing::add () {
     }
     //first is afs and second is customer
     if (f1 != matheus_model.afssF0Indexes.end() && c2 != matheus_model.customersC0Indexes.end()) 
-      for (size_t i = 0; i < matheus_model.c0.size(); ++i) {
+      for (int i = 0; i < matheus_model.c0.size(); ++i) {
         matheus_model.model.add(matheus_model.y[i][f1->second][c2->second] == 0);
         ++matheus_model.nPreprocessings1;
       }
     //first is customer and second is afs 
     if (c1 != matheus_model.customersC0Indexes.end() && f2 != matheus_model.afssF0Indexes.end()) 
-      for (size_t i = 0; i < matheus_model.c0.size(); ++i) {
+      for (int i = 0; i < matheus_model.c0.size(); ++i) {
         matheus_model.model.add(matheus_model.y[c1->second][f2->second][i] == 0);
         ++matheus_model.nPreprocessings1;
       }
     //first is an afs and second is afs 
     if (f1 != matheus_model.afssF0Indexes.end() && f2 != matheus_model.afssF0Indexes.end()) 
-      for (size_t i_ = 0; i_ < matheus_model.c0.size(); ++i_)
-        for (size_t j_ = 0; j_ < matheus_model.c0.size(); ++j_)
-          for (size_t k = 0; k < matheus_model.c0.size(); ++k)
+      for (int i_ = 0; i_ < matheus_model.c0.size(); ++i_)
+        for (int j_ = 0; j_ < matheus_model.c0.size(); ++j_)
+          for (int k = 0; k < matheus_model.c0.size(); ++k)
             matheus_model.model.add(matheus_model.y[i_][f1->second][j_] + matheus_model.y[j_][f2->second][k] <= 1);
   }
 }
