+++ src/models/gvrp_models/cplex/matheus_model/subcycle_user_constraint.cpp
@@ -45,7 +45,7 @@ void Subcycle_user_constraint::main() {
 //    return;
   }
   //setup
-  const size_t sc0 = matheus_model.c0.size(),
+  const int sc0 = matheus_model.c0.size(),
         sf0 = matheus_model.f0.size();
   int bppNRoutesLB, 
       improvedMSTNRoutesLB, 
@@ -63,17 +63,17 @@ void Subcycle_user_constraint::main() {
   //get values
   Matrix2DVal x_vals (env, sc0);
   Matrix3DVal y_vals (env, sc0);
-  for (size_t i = 0; i < sc0; ++i) {
+  for (int i = 0; i < sc0; ++i) {
     x_vals[i] = IloNumArray (env, sc0, 0, 1, IloNumVar::Float);
     y_vals[i] = Matrix2DVal (env, sc0);
     getValues(x_vals[i], matheus_model.x[i]);
-    for (size_t f = 0; f < sf0; ++f) {
+    for (int f = 0; f < sf0; ++f) {
       y_vals[i][f] = IloNumArray (env, sc0, 0, 1, IloNumVar::Float);
       getValues(y_vals[i][f], matheus_model.y[i][f]);
     }
   }
   //creating nodes
-  for (size_t i = 0; i < sc0; ++i) 
+  for (int i = 0; i < sc0; ++i) 
     nodes[i] = graph.addNode();
   //get components
   for (int i = 1; i < sc0; ++i) {
@@ -101,7 +101,7 @@ void Subcycle_user_constraint::main() {
           flowExists = true;
           x_vals[j][curr] = 0;
         }
-        for (size_t f = 0; f < sf0; ++f) {
+        for (int f = 0; f < sf0; ++f) {
           if (y_vals[curr][f][j] > EPS) {
             currToJFlow += y_vals[curr][f][j];
             flowExists = true;
@@ -160,15 +160,15 @@ void Subcycle_user_constraint::main() {
     if (!S.count(0)) {
       //\sum_{v_i \in V'\S} \sum_{v_j \in S} x_{ij} + \sum_{v_f \in F_0} y_{ifj} \geqslant 1 
       //lhs
-      for (size_t i = 0; i < sc0; ++i) 
+      for (int i = 0; i < sc0; ++i) 
         if (!S.count(i)) 
           for (int j : S) {
             lhs += matheus_model.x[i][j];
-            for (size_t f = 0; f < sf0; ++f)
+            for (int f = 0; f < sf0; ++f)
               lhs += matheus_model.y[i][f][j];
           }
       //rhs
-      const size_t sS = S.size();
+      const int sS = S.size();
       vector<const Vertex *> vertices (sS + 1);
       int j = 0;
       for (int i : S) {
@@ -200,9 +200,9 @@ void Subcycle_user_constraint::main() {
       lhs = IloExpr(env);
     }
   //clean
-  for (size_t i = 0; i < sc0; ++i) {
+  for (int i = 0; i < sc0; ++i) {
     x_vals[i].end();
-    for (size_t f = 0; f < sf0; ++f) 
+    for (int f = 0; f < sf0; ++f) 
       y_vals[i][f].end();
     y_vals[i].end();
   }
